Implement "Recently Added" view
There's an endpoint in the Apple Music API for retrieving up to 150 recently added items, it just isn't exposed through its own method in MusicKit.

https://developer.apple.com/documentation/applemusicapi/get_recently_added

The component is almost identical to the albums component, could really do with sharing a component which just takes in a fetch method.

The max limit for fetching is 10 rather than 100 for this method.

Fixes #14