Display song actions menu on right click for the whole table row (#53)
Easier to hit than a small click target on the right.

Ideally it would pop up at the click position, but I'm having trouble getting Bootstrap Vue to do what I want. This is an improvement for now.

Had to hack it a bit to store the ID of the song in the DOM, on each TD element. I've filed a pull request to add contextmenu event support to the table component natively (https://github.com/bootstrap-vue/bootstrap-vue/pull/2064), but this is the only way I could think to do it for now.